third_party/blink/web_tests/http/tests/inspector-protocol/network/interception-take-stream-expected.txt

Tests that resource bodies can be taken as streams.

Running test: testBasicUsage
data: The_quick_brown_fox_jumps_over_the_lazy_dog_0123456789The_quick_brown_fox_jumps_over_the_lazy_dog_01 (100) eof: false
eof: true

Running test: testLargeRead
data: (100000) eof: false
eof: true

Running test: testSmallNonOverlappingReads
data: (10000)

Running test: testSmallOverlappingReads
data: (100000)

Running test: testRead0
reading 0: Invalid max read size
data: The_quick_brown_fox_jumps_over_the_lazy_dog_0123456789The_quick_brown_fox_jumps_over_the_lazy_dog_01 (100) eof: false

Running test: testTakeTwice
Trying to take stream twice: Can only get response body on requests captured after headers received.

Running test: testReadAfterCancel
data: The_quick_brown_fox_jumps_over_the_lazy_dog_0123456789The_quick_brown_fox_jumps_over_the_lazy_dog_01 (100) eof: false
eof: true

Running test: testContinueAfterBodyTaken
Attempting to continue as is after taking request: Unable to continue request as is after body is taken

Running test: testReadAfterClose
Attempting read after close: Invalid stream handle
